Refactor `fake_clock.go` to the `timetesting` package
