+++ internal/timetesting/fake_clock.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package clock
+package timetesting
 
 import (
 	"time"
